* chore(common): Invalid date in budget planing [GCP-11717] (#8646)

* chore(general ledger): Report issues fixed [GCP-11812] (#8649)

* chore(budget): issue on portal download fixed [GCP-11715] (#8652)


* chore(general ledger): Mentioned issues fixed [GCP-11812] (#8653)

* chore(general ledger): Column type altered [GCP-11812] (#8654)

* chore(general ledger): Minor changes in delete [GCP-11817] (#8656)

* feat(budget planning): Budget Planning Details Configuration [GCP-10296] (#8655)

* chore(common): Zero Budget Validation not checked in system [GCP-11746] (#8648)

* chore(budget planning): budget planning status changes (#8657)


* chore(budget planning): budget planning date picker changes [GCP-11737] (#8661)


* chore(general ledger): Template validation added [GCP-11868] (#8663)

* feat(treasury management): b2B file format double quotation mark replaced with empty [GCP-11669] (#8664)

* feat(treasury management): b2B file format double quotation mark replaced with empty [GCP-11669]

* changes

---------


* chore(budget planning): issue on request code duplication fixes [GCP-11822] (#8666)

* chore(budget planning): issue on request code duplication fixes [GCP-11822]

* changesw

---------


* chore(budget): changes on request code [GCP-11822] (#8667)

* chore(budget planning): issue on request code duplication fixes [GCP-11822]

* changesw

* chore(budget): changes on request code [GCP-11822]

---------


* chore(budget): validaiton issue fixed onrequest code [GCP-11822] (#8668)

* chore(budget planning): issue on request code duplication fixes [GCP-11822]

* changesw

* chore(budget): changes on request code [GCP-11822]

* chore(budget): validaiton issue fixed onrequest code [GCP-11822]

---------


* chore(budget): validaiton issue fixed onrequest code [GCP-11822] (#8669)

* chore(budget planning): issue on request code duplication fixes [GCP-11822]

* changesw

* chore(budget): changes on request code [GCP-11822]

* chore(budget): validaiton issue fixed onrequest code [GCP-11822]

* changes

---------


* fix(sales & marketing): fixed error exception [GCP-11711] (#8665)

* chore(common): Error Message Displayed as Unable to update [GCP-11719] (#8662)

* chore(common): Error Message Displayed as Unab

* chore(common): Error Message Displayed as Unab

* chore(budget): budget request date date format changs [GCP-11822] (#8671)


* feat(budget planning): unit and unit item dropdown value changes [GCP-11874] (#8673)


* feat(treasury management): GPOS records in bank reconcilation [GCP-11871] (#8674)

* chore(treasury management): Mentioned defects fixed [GCP-11970] (#8675)

---------

+++ app/Http/Controllers/API/FinalReturnIncomeTemplateDetailsAPIController.php
@@ -10,6 +10,7 @@ use Illuminate\Http\Request;
 use App\Http\Controllers\AppBaseController;
 use App\Models\FinalReturnIncomeTemplateColumns;
 use App\Models\FinalReturnIncomeTemplateDefaults;
+use App\Models\FinalReturnIncomeReports;
 use App\Models\FinalReturnIncomeTemplateLinks;
 use InfyOm\Generator\Criteria\LimitOffsetCriteria;
 use Prettus\Repository\Criteria\RequestCriteria;
@@ -120,6 +121,13 @@ class FinalReturnIncomeTemplateDetailsAPIController extends AppBaseController
     {
         $input = $request->all();
         $input = $this->convertArrayToValue($input);
+
+        $isTemplateUsed = FinalReturnIncomeReports::isTemplateUsed($input['templateMasterID']);
+
+        if($isTemplateUsed) {
+            return $this->sendError('Template already used in a report and cannot be modified', 500);
+        }
+       
         DB::beginTransaction();
         try {
             $validator = \Validator::make($request->all(), [
@@ -272,9 +280,14 @@ class FinalReturnIncomeTemplateDetailsAPIController extends AppBaseController
 
         /** @var FinalReturnIncomeTemplateDetails $finalReturnIncomeTemplateDetails */
         $finalReturnIncomeTemplateDetails = $this->finalReturnIncomeTemplateDetailsRepository->findWithoutFail($id);
+        $isTemplateUsed = FinalReturnIncomeReports::isTemplateUsed($finalReturnIncomeTemplateDetails->templateMasterID);
 
         if (empty($finalReturnIncomeTemplateDetails)) {
             return $this->sendError('Final Return Income Template Details not found');
+        }
+
+        if($isTemplateUsed) {
+            return $this->sendError('Template already used in a report and cannot be updated', 500);
         }
 
         $finalReturnIncomeTemplateDetails = $this->finalReturnIncomeTemplateDetailsRepository->update($input, $id);
@@ -326,13 +339,19 @@ class FinalReturnIncomeTemplateDetailsAPIController extends AppBaseController
         /** @var FinalReturnIncomeTemplateDetails $finalReturnIncomeTemplateDetails */
         $finalReturnIncomeTemplateDetails = $this->finalReturnIncomeTemplateDetailsRepository->findWithoutFail($id);
 
+        $templateMasterID = $finalReturnIncomeTemplateDetails->templateMasterID;
+        $masterID         = $finalReturnIncomeTemplateDetails->masterID;
+        $companySystemID  = $finalReturnIncomeTemplateDetails->companySystemID;
+        $isTemplateUsed = FinalReturnIncomeReports::isTemplateUsed($templateMasterID);
+
         if (empty($finalReturnIncomeTemplateDetails)) {
             return $this->sendError('Final Return Income Template Details not found');
         }
 
-        $templateMasterID = $finalReturnIncomeTemplateDetails->templateMasterID;
-        $masterID         = $finalReturnIncomeTemplateDetails->masterID;
-        $companySystemID  = $finalReturnIncomeTemplateDetails->companySystemID;
+        if($isTemplateUsed) {
+            return $this->sendError('Template already used in a report and cannot be deleted', 500);
+        }
+       
 
         $finalReturnIncomeTemplateDetails->delete();
 
@@ -412,11 +431,13 @@ class FinalReturnIncomeTemplateDetailsAPIController extends AppBaseController
 
             $companySystemID = $request->query('companySystemID');
             $localCurrency = \Helper::companyCurrency($companySystemID);
+            $isTemplateUsed = FinalReturnIncomeReports::isTemplateUsed($templateId);
 
             $output = [
                 'templateDetails' => $templateDetails->toArray(), 
                 'columns' => $templateColumns->toArray(),
-                'localCurrency' => $localCurrency->localcurrency->CurrencyCode
+                'localCurrency' => $localCurrency->localcurrency->CurrencyCode,
+                'isTemplateUsed' => $isTemplateUsed
             ];
 
         return $this->sendResponse($output, 'Final Return Income Template Details retrieved successfully');
@@ -424,6 +445,13 @@ class FinalReturnIncomeTemplateDetailsAPIController extends AppBaseController
 
     public function templateDetailRaw(Request $request) {
         $input = $request->all();
+
+        $isTemplateUsed = FinalReturnIncomeReports::isTemplateUsed($input['templateMasterID']);
+
+        if($isTemplateUsed) {
+            return $this->sendError('Template already used in a report and cannot be modified', 500);
+        }
+       
 
         $maxSortOrder = FinalReturnIncomeTemplateDetails::where('templateMasterID', $input['templateMasterID'])
             ->where('masterID', $input['masterID'])
